Refactor saml auth functions into separate php partial
